angle: split angle.rs from rotation.rs
